+++ client/examples/output/build_basic_output.rs
@@ -32,7 +32,7 @@ async fn main() -> Result<()> {
 
     let token_supply = client.get_token_supply().await?;
 
-    let address = Address::try_from_bech32("rms1qpllaj0pyveqfkwxmnngz2c488hfdtmfrj3wfkgxtk4gtyrax0jaxzt70zy")?.1;
+    let address = Address::try_from_bech32("rms1qpllaj0pyveqfkwxmnngz2c488hfdtmfrj3wfkgxtk4gtyrax0jaxzt70zy")?;
 
     let basic_output_builder = BasicOutputBuilder::new_with_amount(1_000_000)?
         .add_unlock_condition(UnlockCondition::Address(AddressUnlockCondition::new(address)));